move mobile dropdown into own component
